fix(architecture): extract Application/Actions for Cameras, EdgeSync, Billing trial (#1411)
api/ARCHITECTURE.md flagged Cameras (0 Actions / 6 controllers) and
EdgeSync (0 Actions / 3 controllers) as having zero use-case extraction,
with all business logic living directly in controllers/services,
contrary to the documented 1 Action = 1 use case pattern.

Extracted the write-path use cases:
- Cameras\Application\Actions\{CreateCamera,UpdateCamera,DeleteCamera}
- EdgeSync\Application\Actions\{RegisterEdgeNode,PushEdgeRecords}
- Billing\Application\Actions\{RequestTrialSignup,VerifyTrialSignup}
  (splits SelfServiceTrialController, the 531-line example explicitly
  cited in the ticket)

Controllers now inject and delegate to these Actions instead of
inlining DB/service calls; SelfServiceTrialController::signup()/verify()
are pure orchestrators (validate -> Action -> JSON response).

Pure refactor, no behavior change.

Verified:
- php -l clean on all 10 touched/new files
- vendor/bin/pint --test clean
- Targeted PHPStan: 45 errors vs 57 pre-existing on the same files
  (net reduction, zero new findings; remainder tracked as #1413)
- Cameras feature tests: 21/21 green
- Edge/EdgeSync feature+integration tests: all green, 0 failures
  (real local Postgres 16 test database)

Fixes #1411
