+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## [Unreleased]
 
 ### Fixed
+- **fix(architecture): extract Application/Actions for Cameras, EdgeSync, and the Billing self-service trial flow (issue #1411).** `api/ARCHITECTURE.md` already flagged "trop peu d'Actions, controllers trop epais" and an audit confirmed `Cameras` (0 Actions / 6 controllers) and `EdgeSync` (0 Actions / 3 controllers) had zero use-case extraction — all business logic lived directly in controllers/services, contrary to the documented "1 Action = 1 use case" pattern (`CONVENTIONS.md` §2.7). Extracted the write-path use cases: `Cameras\Application\Actions\{CreateCamera,UpdateCamera,DeleteCamera}` (each a thin wrapper delegating to the existing `CameraService`, injected into `CameraController` instead of `new`); `EdgeSync\Application\Actions\{RegisterEdgeNode,PushEdgeRecords}` (node registration + license issuance, and Edge daemon push-to-queue + job dispatch, previously inlined in `EdgeNodeController::store()`/`pushFromEdge()`). Also tackled the ticket's other named target, `SelfServiceTrialController` (531 lines, DB/Mail/Hash calls directly in the controller, no Action at all): split into `Billing\Application\Actions\RequestTrialSignup` (OTP generation, pending `CompanyRequest` creation, verification email) and `VerifyTrialSignup` (OTP check, trial `Company`+`Employee` provisioning transaction, welcome email, drip-email scheduling) — the controller's `signup()`/`verify()` methods are now pure orchestrators (validate → Action → JSON response), matching the pattern already used by `HR\Application\Actions\CreateEmployee`. Pure refactor, no behavior change: `EdgeSync`/`Payroll`/`HR` modules still at their pre-existing Actions count for the modules not touched here (tracked separately if a full sweep across `HR`/`Payroll`/`Platform` is wanted, per the ticket's "peut etre split en sous-issues" note). Verified: `php -l` clean on all 10 touched/new files; `vendor/bin/pint --test` clean; targeted PHPStan run on the same 10 files shows 45 pre-existing errors vs. 57 on the 3 original files before the split (net reduction, zero new findings, remainder tracked separately as #1413); full Cameras (21/21) and Edge/EdgeSync (Feature+Integration, 0 failures) test suites green against a real local Postgres 16 test database — no regression on create/update/delete camera flows, Edge node registration, or push/pull/heartbeat sync.
 - **fix(critical): resolve 3 high-priority issues — EdgeSync duplicate migration (#1394), Billing/Payroll circular Domain dependency (#1395), CI untrusted-checkout CodeQL alert (#1396).** (1) `api/app/Modules/EdgeSync/database/migrations/2026_06_29_000001_create_edge_sync_tables.php` (module-local copy with strict non-nullable FK on `company_id`, incompatible with shared-tenant PostgreSQL on Render) deleted; `EdgeSyncServiceProvider::loadMigrationsFrom()` removed — EdgeSync migrations now live exclusively in `api/database/migrations/tenant/` like all other modules, eliminating the `migrate:fresh` duplicate-table collision. (2) Circular compile-time `use` imports between `Billing\Domain\Models\Invoice`↔`Payroll\Domain\Models\Payment` and `Billing\Domain\Models\Partner`↔`Payroll\Domain\Models\Commission` broken: the two Billing models now reference Payroll classes via FQCN strings in `hasMany()` — no static import at all, Eloquent resolves at runtime, PHPDoc `@return` retains the full class name so PHPStan/IDEs keep type inference. Natural direction (Payroll→Billing) preserved unchanged. ADR-0011 documents the decision. (3) CodeQL `actions/untrusted-checkout/critical` Alert #29 (Option A fix): `distribute-mobile` job extracted from `deploy-main.yml` (which runs in a `workflow_run` privileged context) into a new dedicated workflow `.github/workflows/mobile-distribute-main.yml` triggered by `push: [main]` scoped to `front/mobile_apps/**` — `github.sha` is always a verified commit on the protected branch, eliminating the untrusted-checkout class entirely. A no-op stub job retained in `deploy-main.yml` preserves any existing branch-protection required-status-checks during the transition period. New workflow supports `workflow_dispatch` with an optional `app` input (employee|manager|platform_admin|all) for targeted manual re-distribution.
 - **Docs audit P2: stale `front/mobile/` path in `docs/GESTION_PROJET/DOSSIER_REPONSE_AU_CAHIER_DES_CHARGES.md`.** This 2026-05-10 status document described the mobile app as living in `front/mobile/`, a path removed from the repo on 2026-06-13 (PR #754) and replaced by 5 separate Flutter apps under `front/mobile_apps/` (`leopardo_core`, `leopardo_employee`, `leopardo_hr`, `leopardo_manager`, `leopardo_platform_admin`). A dev agent following this document literally would look for a directory that no longer exists. Added a dated freshness banner at the top of the file (documenting the drift explicitly, pointing to `PILOTAGE.md` for current product state, consistent with the existing banner convention on `GARDE_FOUS.md`) and corrected both `front/mobile/` occurrences (product overview list, mobile CI conditional-trigger bullet) to `front/mobile_apps/`, while preserving the original wording as historical context. No code change.
